+++ src/3d/qgs3dutils.h
@@ -355,10 +355,14 @@ class _3D_EXPORT Qgs3DUtils
 
     /**
      * Transform the given screen point to \a QgsPoint in map coordinates
-     *
+     * \note This transformation is not unique as we pick point halfway between near and far plane on the ray from the camera.
+     * \param size size of canvas
+     * \param screenPoint point in screen coordinates
+     * \param cameraController camera controller attached to the scene
+     * \param mapSettings 3D map settings of the scene
      * \since QGIS 3.44
      */
-    static QgsPoint screenPointToMapCoordinates( const QPoint &screenPoint, Qgs3DMapCanvas &canvas );
+    static QgsPoint screenPointToMapCoordinates( const QPoint &screenPoint, QSize size, const QgsCameraController *cameraController, const Qgs3DMapSettings *mapSettings );
 };
 
 #endif // QGS3DUTILS_H
